+++ sec5.config.js
@@ -1,6 +1,6 @@
 export default {
   /**
-   * Allowed beacuse of `src/todoCheck.ts` and build (`bin/sec5.js`).
+   * Allowed beacuse of `src/check.ts` and build (`bin/sec5.js`) includes strings like FIXME and more.
    * The number of comments should not exceed 2.
    */
   allowComments: true,