+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2]
+
+### Added
+
+- `Eslint` check
+- Information about `ESlint` check
+
+### Fixed
+
+- `Prettier` check doesn't exit and close the process
+
 ## [0.1.1]
 
 ### Added